acf/vss: fix unbounded reads in string-array (de)serialisation
Avtp_Vss_DeserializeStringArray and Avtp_Vss_GetVSSDataStringArrayLength
in src/avtp/acf/custom/Vss.c walk an on-wire string-array layout of
[length-prefix : 2 bytes][string : length bytes] pairs.  Both had
serious bounds-check defects allowing crafted network input to drive
out-of-bounds reads:

1. Avtp_Vss_DeserializeStringArray declared `uint16_t idx = 0;` and
   only checked `if (idx >= array_length) break;` — but `idx` was
   never incremented, so the break was dead code (it only fired when
   array_length == 0).  With attacker-controlled str_length values
   from the wire (uint16_t, up to 65535), the memcpy() inside the
   loop could read up to ~64 KiB past the end of
   vss_data_string_array->data on every iteration after the first
   malicious entry.  Heap-disclosure / crash class.

2. Avtp_Vss_GetVSSDataStringArrayLength advanced
   `ptr_idx += 2 + str_length` in uint16_t arithmetic.  A str_length
   value near UINT16_MAX wraps ptr_idx around to a small number,
   making the loop iterate from a low offset again.  Pseudo-infinite
   loop / DoS class.  In addition, the loop condition
   `ptr_idx < total_length` was off by one — it permitted reading the
   2-byte length prefix when only 1 byte remained.

3. Avtp_Vss_GetVSSDataStringArrayLength also returned uint8_t for a
   count that could plausibly exceed 255 (since total_length is
   uint16_t and a string can be 0 bytes, the protocol allows up to
   ~32k strings in a single array).  Same shape as
   Avtp_Can_GetCanPayloadLength fixed in the previous PR.

Fixes:

- Add a `consumed` counter in DeserializeStringArray that is actually
  incremented per iteration and used to drive the break.
- Add an explicit "remaining bytes >= str_length" check before
  memcpy() so a single malformed string-length value cannot trigger
  the OOB read.
- In GetVSSDataStringArrayLength, change the loop guard to
  `ptr_idx + 2 <= total_length` (uint32_t arithmetic), and add a
  matching post-length-read check that the string body fits.
- Widen the GetVSSDataStringArrayLength return type and the function
  declaration in include/avtp/acf/custom/Vss.h from uint8_t to
  uint16_t.
- Add a NULL-check for strings[i] before dereferencing it (the
  previous code would NULL-deref strings[i]->data_length on the
  first malformed entry of an under-allocated output array).

All bound-check additions are computed in uint32_t so the additions
themselves cannot overflow when str_length approaches UINT16_MAX.

No behaviour change for well-formed input.  Existing unit tests in
unit/test-vss.c continue to pass (they exercise the 3-string happy
path).

+++ include/avtp/acf/custom/Vss.h
@@ -258,7 +258,7 @@ Vss_Datatype_t Avtp_Vss_GetDatatype(const Avtp_Vss_t* const pdu);
 uint64_t Avtp_Vss_GetMsgTimestamp(const Avtp_Vss_t* const pdu);
 void Avtp_Vss_GetVssPath(const Avtp_Vss_t* const pdu, VssPath_t* val);
 void Avtp_Vss_GetVssData(const Avtp_Vss_t* const pdu, VssData_t* val);
-uint8_t Avtp_Vss_GetVSSDataStringArrayLength(const VssDataStringArray_t* str_array);
+uint16_t Avtp_Vss_GetVSSDataStringArrayLength(const VssDataStringArray_t* str_array);
 uint16_t Avtp_Vss_CalcVssPathLength (const Avtp_Vss_t* const pdu);
 void Avtp_Vss_DeserializeStringArray(const VssDataStringArray_t*  const vss_data_string_array,
                                      VssDataString_t* strings[],
